move test to proxy pkg
